Update substrate, evm and ethereum versions

+++ frame/evm/precompile/bn128/src/lib.rs
@@ -46,12 +46,12 @@ fn read_point(input: &[u8], start_inx: usize) -> Result<bn::G1, ExitError> {
 pub struct Bn128Add;
 
 impl LinearCostPrecompile for Bn128Add {
-	const BASE: usize = 15;
-	const WORD: usize = 3;
+	const BASE: u64 = 15;
+	const WORD: u64 = 3;
 
 	fn execute(
 		input: &[u8],
-		_: usize,
+		_: u64,
 	) -> core::result::Result<(ExitSucceed, Vec<u8>), ExitError> {
 		use bn::AffineG1;
 
@@ -73,12 +73,12 @@ impl LinearCostPrecompile for Bn128Add {
 pub struct Bn128Mul;
 
 impl LinearCostPrecompile for Bn128Mul {
-	const BASE: usize = 15;
-	const WORD: usize = 3;
+	const BASE: u64 = 15;
+	const WORD: u64 = 3;
 
 	fn execute(
 		input: &[u8],
-		_: usize,
+		_: u64,
 	) -> core::result::Result<(ExitSucceed, Vec<u8>), ExitError> {
 		use bn::AffineG1;
 
@@ -100,12 +100,12 @@ impl LinearCostPrecompile for Bn128Mul {
 pub struct Bn128Pairing;
 
 impl LinearCostPrecompile for Bn128Pairing {
-	const BASE: usize = 15;
-	const WORD: usize = 3;
+	const BASE: u64 = 15;
+	const WORD: u64 = 3;
 
 	fn execute(
 		input: &[u8],
-		_: usize,
+		_: u64,
 	) -> core::result::Result<(ExitSucceed, Vec<u8>), ExitError> {
 		use bn::{AffineG1, AffineG2, Fq, Fq2, pairing_batch, G1, G2, Gt, Group};
